case2/requirements.txt
horovod
psutil
py-cpuinfo
pynvml
tensorflow
scikit-learn
scikit-image
numpy
scipy
h5py
click
matplotlib
pytest
tqdm
xarray
h5netcdf
tables
pandas
